requirements.txt
asgiref==3.8.1
awscli==1.33.21
boto3==1.34.139
botocore==1.34.139
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docutils==0.16
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.4.0
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.10.2
setuptools==70.2.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.2
whitenoise==6.7.0
